+++ javascript/speechbubble/displayObject.js
