+++ javascript/speechbubble/interactiveObject.js
